Split conf, skip reply
- Split conf structs and functions for file, interface and client_pool
  configuration to separate files
- Add client_pool option to skip waiting for any reply
